Revert "Implement new version schema for cosmosdb starter (#14465)"
This reverts commit a5acccf35cbf14ce990188b4f60920bc7f3a6292.
