+++ eng/pipelines/templates/stages/cosmos-sdk-client.yml
@@ -153,9 +153,9 @@ stages:
           - script: |
               python --version
               python -m pip install setuptools termcolor in_place
-              python sdk/spring/scripts/replace_util.py --module cosmos --log debug --color false
+              python sdk/spring/scripts/replace_artifact.py --module cosmos --log debug --color false
             condition: eq(variables['DisplayName'], 'Spring Emulator only Integration Tests')
-            displayName: 'Run replace_util.py'
+            displayName: 'Run replace_artifact.py'
           - task: Maven@3
             condition: eq(variables['DisplayName'], 'Spring Emulator only Integration Tests')
             displayName: 'Run tests again: $(DisplayName)'
