+++ eng/versioning/version_client.txt
@@ -91,8 +91,7 @@ com.azure.resourcemanager:azure-resourcemanager-trafficmanager;2.0.0-beta.4;2.0.
 com.azure.resourcemanager:azure-resourcemanager-test;2.0.0-beta.4;2.0.0-beta.5
 com.microsoft.azure:azure-active-directory-b2c-spring-boot-starter;2.3.3;2.4.0-beta.1
 com.microsoft.azure:azure-active-directory-spring-boot-starter;2.3.3;2.4.0-beta.1
-com.microsoft.azure:azure-cosmosdb-spring-boot-2-3-starter;2.4.0-beta.1;2.4.0-beta.1
-com.microsoft.azure:azure-cosmosdb-spring-boot-2-2-starter;2.4.0-beta.1;2.4.0-beta.1
+com.microsoft.azure:azure-cosmosdb-spring-boot-starter;2.3.3;2.4.0-beta.1
 com.microsoft.azure:azure-data-gremlin-spring-boot-starter;2.3.3;2.4.0-beta.1
 com.microsoft.azure:azure-keyvault-secrets-spring-boot-starter;2.3.3;2.4.0-beta.1
 com.microsoft.azure:azure-servicebus-jms-spring-boot-starter;2.3.3;2.4.0-beta.1
