--- sdk/spring/azure-spring-boot-2-2-starter-cosmosdb/CHANGELOG.md
@@ -1,4 +1,0 @@
-# Release History
-
-
-## 2.4.0-beta.1 (Unreleased)
