--- sdk/spring/azure-spring-boot-2-3-starter-cosmosdb/src/main/resources/cosmosdb.enable.config
@@ -1,1 +1,0 @@
-dummy
