+++ sdk/spring/azure-spring-boot-starter-cosmosdb/README.md
@@ -1,4 +1,4 @@
-# Azure Cosmos DB Spring Boot 2.3 Starter client library for Java
+# Azure Cosmos DB Spring Boot Starter client library for Java
 
 [Azure Cosmos DB](https://azure.microsoft.com/services/cosmos-db/) is a globally-distributed database service that allows developers to work with data using a variety of standard APIs, such as SQL, MongoDB, Graph, and Azure Table storage. 
 
@@ -11,12 +11,12 @@
 - [Maven](http://maven.apache.org/) 3.0 and above
 
 ### Include the package
-[//]: # ({x-version-update-start;com.microsoft.azure:azure-cosmosdb-spring-boot-2-3-starter;current})
+[//]: # ({x-version-update-start;com.microsoft.azure:azure-cosmosdb-spring-boot-starter;current})
 ```xml
 <dependency>
     <groupId>com.microsoft.azure</groupId>
-    <artifactId>azure-cosmosdb-spring-boot-2-3-starter</artifactId>
-    <version>2.4.0-beta.1</version>
+    <artifactId>azure-cosmosdb-spring-boot-starter</artifactId>
+    <version>2.3.3</version>
 </dependency>
 ```
 [//]: # ({x-version-update-end})
@@ -235,12 +235,12 @@ Besides using this Azure CosmosDb Spring Boot Starter, you can directly use Spri
 
 The following section provide a sample project illustrating how to use the starter.
 ### More sample code
-- [Cosmos DB SQL API][cosmos_db_sql_api]
+- [Cosmos DB SQL API](../azure-spring-boot-samples/azure-spring-boot-sample-cosmosdb)
 
 ## Contributing
 This project welcomes contributions and suggestions.  Most contributions require you to agree to a Contributor License Agreement (CLA) declaring that you have the right to, and actually do, grant us the rights to use your contribution. For details, visit https://cla.microsoft.com.
 
-Please follow [instructions here][contributing_readme] to build from source or contribute.
+Please follow [instructions here](../CONTRIBUTING.md) to build from source or contribute.
 
 <!-- LINKS -->
 [docs]: https://docs.microsoft.com/azure/developer/java/spring-framework/configure-spring-boot-starter-java-app-with-cosmos-db
@@ -249,5 +249,3 @@ Please follow [instructions here][contributing_readme] to build from source or c
 [sample]: https://github.com/Azure/azure-sdk-for-java/tree/master/sdk/spring/azure-spring-boot-samples/azure-spring-boot-sample-cosmosdb
 [logging]: https://github.com/Azure/azure-sdk-for-java/wiki/Logging-with-Azure-SDK#use-logback-logging-framework-in-a-spring-boot-application
 [azure_subscription]: https://azure.microsoft.com/free
-[cosmos_db_sql_api]: https://github.com/Azure/azure-sdk-for-java/tree/master/sdk/spring/azure-spring-boot-samples/azure-spring-boot-sample-cosmosdb
-[contributing_readme]: https://github.com/Azure/azure-sdk-for-java/blob/master/sdk/spring/CONTRIBUTING.md
