+++ sdk/spring/azure-spring-boot-starter-cosmosdb/pom.xml
@@ -12,11 +12,11 @@
   </parent>
 
   <groupId>com.microsoft.azure</groupId>
-  <artifactId>azure-cosmosdb-spring-boot-2-3-starter</artifactId>
-  <version>2.4.0-beta.1</version> <!-- {x-version-update;com.microsoft.azure:azure-cosmosdb-spring-boot-2-3-starter;current} -->
+  <artifactId>azure-cosmosdb-spring-boot-starter</artifactId>
+  <version>2.4.0-beta.1</version> <!-- {x-version-update;com.microsoft.azure:azure-cosmosdb-spring-boot-starter;current} -->
 
-  <name>Azure Cosmos DB Spring Boot 2.3 Starter</name>
-  <description>Spring Boot 2.3 Starter for Azure Cosmos DB service</description>
+  <name>Azure Cosmos DB Spring Boot Starter</name>
+  <description>Spring Boot Starter for Azure Document DB service</description>
   <url>https://github.com/Azure/azure-sdk-for-java</url>
 
   <dependencies>
