+++ sdk/spring/azure-spring-boot-starter-cosmosdb/src/main/resources/cosmosdb.enable.config
