+++ sdk/spring/azure-spring-boot-test-cosmosdb/pom.xml
@@ -17,8 +17,8 @@
   <dependencies>
     <dependency>
       <groupId>com.microsoft.azure</groupId>
-      <artifactId>azure-cosmosdb-spring-boot-2-3-starter</artifactId>
-      <version>2.4.0-beta.1</version> <!-- {x-version-update;com.microsoft.azure:azure-cosmosdb-spring-boot-2-3-starter;current} -->
+      <artifactId>azure-cosmosdb-spring-boot-starter</artifactId>
+      <version>2.4.0-beta.1</version> <!-- {x-version-update;com.microsoft.azure:azure-cosmosdb-spring-boot-starter;current} -->
     </dependency>
     <dependency>
       <groupId>com.azure</groupId>
@@ -57,7 +57,7 @@
           <rules>
             <bannedDependencies>
               <includes>
-                <include>com.microsoft.azure:azure-cosmosdb-spring-boot-2-3-starter:[2.4.0-beta.1]</include> <!-- {x-include-update;com.microsoft.azure:azure-cosmosdb-spring-boot-2-3-starter;current} -->
+                <include>com.microsoft.azure:azure-cosmosdb-spring-boot-starter:[2.4.0-beta.1]</include> <!-- {x-include-update;com.microsoft.azure:azure-cosmosdb-spring-boot-starter;current} -->
                 <include>org.springframework.boot:spring-boot-starter-web:[2.3.3.RELEASE]</include> <!-- {x-include-update;org.springframework.boot:spring-boot-starter-web;external_dependency} -->
                 <include>org.springframework.boot:spring-boot-starter-actuator:[2.3.3.RELEASE]</include> <!-- {x-include-update;org.springframework.boot:spring-boot-starter-actuator;external_dependency} -->
               </includes>
