+++ sdk/spring/azure-spring-boot/README.md
@@ -43,8 +43,7 @@ This project provides auto-configuration for the following Azure services:
 
 - [Azure Active Directory](../azure-spring-boot-starter-active-directory)
 - [Azure Active Directory B2C](../azure-spring-boot-starter-active-directory-b2c)
-- [Spring Boot 2.2 for Cosmos DB SQL API](../azure-spring-boot-2-2-starter-cosmosdb)
-- [Spring Boot 2.3 for Cosmos DB SQL API](../azure-spring-boot-2-3-starter-cosmosdb)
+- [Cosmos DB SQL API](../azure-spring-boot-starter-cosmosdb)
 - [Gremlin SQL API](../azure-spring-boot-starter-data-gremlin)
 - [Key Vault Secrets](../azure-spring-boot-starter-keyvault-secrets)
 - [Metrics Service](../azure-spring-boot-starter-metrics)
