+++ sdk/spring/pom.xml
@@ -13,8 +13,7 @@
     <module>azure-spring-boot-starter</module>
     <module>azure-spring-boot-starter-active-directory</module>
     <module>azure-spring-boot-starter-active-directory-b2c</module>
-    <module>azure-spring-boot-2-2-starter-cosmosdb</module>
-    <module>azure-spring-boot-2-3-starter-cosmosdb</module>
+    <module>azure-spring-boot-starter-cosmosdb</module>
     <module>azure-spring-boot-starter-data-gremlin</module>
     <module>azure-spring-boot-starter-keyvault-secrets</module>
     <module>azure-spring-boot-starter-metrics</module>
